daemon: Populate option.Config earlier in initEnv
Some code is beginning to leak in which expects option.Config to be
initialized part way through initEnv(). To ensure that the config is
properly populated with the values from the CLI, and that any references
to the option.Config read the correct value, shift the population of the
config earlier - into initEnv().

Signed-off-by: Joe Stringer <joe@cilium.io>